+++ src/web/static_assets.rs
@@ -6,3 +6,8 @@ pub const APP_JS: &str = include_str!("../../static/app.js");
 pub const MANIFEST_JSON: &str = include_str!("../../static/manifest.json");
 pub const SW_JS: &str = include_str!("../../static/sw.js");
 pub const ICON_SVG: &str = include_str!("../../static/icon.svg");
+
+// Vendored xterm.js (MIT, see static/vendor/LICENSE-xterm.txt) for the pi page.
+pub const XTERM_JS: &str = include_str!("../../static/vendor/xterm.js");
+pub const XTERM_CSS: &str = include_str!("../../static/vendor/xterm.css");
+pub const XTERM_FIT_JS: &str = include_str!("../../static/vendor/xterm-addon-fit.js");